Extract timelock in module
