+++ src/chain/bitcoin/mod.rs
@@ -16,6 +16,7 @@ use crate::role::Arbitrating;
 
 use address::Address;
 use amount::Amount;
+use timelock::CSVTimelock;
 use transaction::{Buy, Cancel, Funding, Lock, Punish, Refund, Tx};
 
 use std::fmt::Debug;
@@ -26,6 +27,7 @@ pub mod address;
 pub mod amount;
 pub mod fee;
 pub mod tasks;
+pub mod timelock;
 pub mod transaction;
 
 #[derive(Clone, Debug, Copy, Eq, PartialEq)]
@@ -74,44 +76,6 @@ impl Timelock for Bitcoin {
 }
 
 impl Arbitrating for Bitcoin {}
-
-impl FromStr for CSVTimelock {
-    type Err = consensus::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let x = s
-            .parse::<u32>()
-            .map_err(|_| consensus::Error::ParseFailed("Failed parsing CSV timelock"))?;
-        Ok(CSVTimelock(x))
-    }
-}
-
-#[derive(PartialEq, Eq, PartialOrd, Clone, Debug, StrictDecode, StrictEncode, Copy)]
-pub struct CSVTimelock(u32);
-
-impl CSVTimelock {
-    pub fn new(timelock: u32) -> Self {
-        Self(timelock)
-    }
-
-    pub fn as_u32(&self) -> u32 {
-        self.0
-    }
-}
-
-impl Encodable for CSVTimelock {
-    fn consensus_encode<W: io::Write>(&self, writer: &mut W) -> Result<usize, io::Error> {
-        bitcoin::consensus::encode::Encodable::consensus_encode(&self.0, writer)
-    }
-}
-
-impl Decodable for CSVTimelock {
-    fn consensus_decode<D: io::Read>(d: &mut D) -> Result<Self, consensus::Error> {
-        let timelock: u32 = bitcoin::consensus::encode::Decodable::consensus_decode(d)
-            .map_err(|_| consensus::Error::ParseFailed("Bitcoin u32 timelock parsing failed"))?;
-        Ok(CSVTimelock(timelock))
-    }
-}
 
 impl Onchain for Bitcoin {
     /// Defines the transaction format used to transfer partial transaction between participant for
