+++ src/chain/bitcoin/timelock.rs
@@ -1,0 +1,45 @@
+use strict_encoding::{StrictDecode, StrictEncode};
+
+use crate::consensus::{self, Decodable, Encodable};
+
+use std::fmt::Debug;
+use std::io;
+use std::str::FromStr;
+
+impl FromStr for CSVTimelock {
+    type Err = consensus::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let x = s
+            .parse::<u32>()
+            .map_err(|_| consensus::Error::ParseFailed("Failed parsing CSV timelock"))?;
+        Ok(CSVTimelock(x))
+    }
+}
+
+#[derive(PartialEq, Eq, PartialOrd, Clone, Debug, StrictDecode, StrictEncode, Copy)]
+pub struct CSVTimelock(u32);
+
+impl CSVTimelock {
+    pub fn new(timelock: u32) -> Self {
+        Self(timelock)
+    }
+
+    pub fn as_u32(&self) -> u32 {
+        self.0
+    }
+}
+
+impl Encodable for CSVTimelock {
+    fn consensus_encode<W: io::Write>(&self, writer: &mut W) -> Result<usize, io::Error> {
+        bitcoin::consensus::encode::Encodable::consensus_encode(&self.0, writer)
+    }
+}
+
+impl Decodable for CSVTimelock {
+    fn consensus_decode<D: io::Read>(d: &mut D) -> Result<Self, consensus::Error> {
+        let timelock: u32 = bitcoin::consensus::encode::Decodable::consensus_decode(d)
+            .map_err(|_| consensus::Error::ParseFailed("Bitcoin u32 timelock parsing failed"))?;
+        Ok(CSVTimelock(timelock))
+    }
+}
